feat: refactor Telegram integration with improved client structure and message handling
