+++ internal/telegram/types.go
@@ -1,0 +1,42 @@
+package telegram
+
+type Message struct {
+	Chat struct {
+		ID        int64  `json:"id"`
+		FirstName string `json:"first_name"`
+	} `json:"chat"`
+	Text string `json:"text"`
+}
+
+type CallbackQuery struct {
+	ID      string   `json:"id"`
+	Data    string   `json:"data"`
+	Message *Message `json:"message"`
+}
+
+type Update struct {
+	UpdateId      int            `json:"update_id"`
+	Message       *Message       `json:"message"`
+	CallbackQuery *CallbackQuery `json:"callback_query"`
+}
+
+type GetUpdatesResponse struct {
+	Ok     bool     `json:"ok"`
+	Result []Update `json:"result"`
+}
+
+// Send message
+type SendMessage struct {
+	ChatId      int64        `json:"TG_CHAT_ID"`
+	Text        string       `json:"text"`
+	ReplyMarkup *ReplyMarkup `json:"reply_markup,omitempty"`
+}
+
+type ReplyMarkup struct {
+	InlineKeyboard [][]InlineKeyboardButton `json:"inline_keyboard"`
+}
+
+type InlineKeyboardButton struct {
+	Text         string `json:"text"`
+	CallbackData string `json:"callback_data"`
+}
